requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cloudinary==1.23.0
dj-database-url==0.5.0
Django==3.1.4
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
